# Significantly faster cohorts detection.