--- a/flox/core.py
+++ b/flox/core.py
@@ -220,12 +220,12 @@ def find_group_cohorts(labels, chunks, merge: bool = True) -> dict:
 
     #  Iterate over each block and create a new block of same shape with "chunk number"
     shape = tuple(array.blocks.shape[ax] for ax in axis)
-    blocks = np.empty(math.prod(shape), dtype=object)
+    blocks = np.empty(shape, dtype=object)
     array_chunks = tuple(np.array(c) for c in array.chunks)
-    for idx, blockindex in enumerate(np.ndindex(array.numblocks)):
+    for blockindex in np.ndindex(array.numblocks):
         chunkshape = get_chunk_shape(array_chunks, blockindex)
-        blocks[idx] = np.full(chunkshape, idx)
-    which_chunk = np.block(blocks.reshape(shape).tolist()).reshape(-1)
+        blocks[blockindex] = np.full(chunkshape, np.ravel_multi_index(blockindex, array.numblocks))
+    which_chunk = np.block(blocks.tolist()).reshape(-1)
 
     raveled = labels.reshape(-1)
     # these are chunks where a label is present