@dcherian 2023-10-10
Down to
```
546 ms ± 3.68 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```