@dcherian 2023-10-11
| Change   | Before [9f82e19d] <main>   | After [fa934069] <faster-cohorts>   |   Ratio | Benchmark (Parameter)                                   |
|----------|----------------------------|-------------------------------------|---------|---------------------------------------------------------|
| -        | 192±1ms                    | 172±4ms                             |    0.9  | cohorts.ERA5DayOfYear.time_graph_construct              |
| -        | 186±3ms                    | 158±2ms                             |    0.85 | cohorts.ERA5DayOfYearRechunked.time_graph_construct     |
| -        | 95.5±0.9ms                 | 68.6±0.9ms                          |    0.72 | cohorts.ERA5DayOfYearRechunked.time_find_group_cohorts  |
| -        | 48.0±0.1ms                 | 25.0±0.1ms                          |    0.52 | cohorts.ERA5DayOfYear.time_find_group_cohorts           |
| -        | 44.5±2ms                   | 18.6±0.6ms                          |    0.42 | cohorts.ERA5MonthHourRechunked.time_graph_construct     |
| -        | 42.3±2ms                   | 17.1±0.2ms                          |    0.4  | cohorts.ERA5MonthHour.time_graph_construct              |
| -        | 9.52±0.04ms                | 3.72±0.01ms                         |    0.39 | cohorts.PerfectMonthly.time_graph_construct             |
| -        | 9.60±0.2ms                 | 3.72±0.02ms                         |    0.39 | cohorts.PerfectMonthlyRechunked.time_graph_construct    |
| -        | 73.5±0.9ms                 | 25.4±0.5ms                          |    0.35 | cohorts.time_cohorts_era5_single                        |
| -        | 31.2±0.2ms                 | 7.58±0.1ms                          |    0.24 | cohorts.ERA5MonthHour.time_find_group_cohorts           |
| -        | 34.1±0.1ms                 | 7.80±0.2ms                          |    0.23 | cohorts.ERA5MonthHourRechunked.time_find_group_cohorts  |
| -        | 6.87±0.03ms                | 1.02±0.04ms                         |    0.15 | cohorts.PerfectMonthlyRechunked.time_find_group_cohorts |
| -        | 6.95±0.1ms                 | 1.00±0.01ms                         |    0.14 | cohorts.PerfectMonthly.time_find_group_cohorts          |